\name{bivrp-internal}
\alias{bp_maxmin}
\alias{budburst_pred}
\alias{exp_dec_wrapper}
\alias{exp_dec}
\alias{get_chill_days}
\alias{get_day_length}
\alias{get_thermal_units}
\alias{get_Tmean}
\alias{lactin_wrapper}
\alias{lactin}
\alias{stage_change_pred_gen2}
\alias{stage_change_pred}
\alias{weibull3p_inverse_wrapper}
\alias{weibull3p_inverse}
\alias{weibull3p}
\encoding{UTF-8}
\title{Internal functions to used to evaluate a phenology model and plot the results}
\description{Internal functions used to evaluate a phenology model and plot the results}
\author{
Rafael A. Moral <rafael.deandrademoral@mu.ie> and Rowan Fealy
}