\name{phenModel-package}
\alias{phenModel-package}
\alias{phenModel}
\docType{package}
\title{
\packageTitle{phenModel}
}
\description{
\packageDescription{phenModel}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{phenModel}
\packageIndices{phenModel}
}
\author{
\packageAuthor{phenModel}

Maintainer: \packageMaintainer{phenModel}
}
\references{
Pollard C.P., Griffin C.T., Moral R.A., Duffy C., Chuche J., Gaffney M.T., Fealy R. (submitted) phenModel: a temperature-dependent phenology model for the blue willow beetle, Phratora vulgatissima.
}